move TestRepository_getAuthMethods to a better location
