+++ internal/auth/oidc/repository_auth_method_read_test.go
@@ -22,13 +22,16 @@ func TestRepository_LookupAuthMethod(t *testing.T) {
 	org, _ := iam.TestScopes(t, iam.TestRepo(t, conn, wrapper))
 	databaseWrapper, err := kmsCache.GetWrapper(context.Background(), org.PublicId, kms.KeyPurposeDatabase)
 	require.NoError(t, err)
-	am := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice.com")[0], "alice_rp", "alices-dogs-name")
+	amInactive := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice-inactive.com")[0], "alice_rp", "alices-dogs-name", WithOperationalState(InactiveState))
+	amActivePriv := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice-active-priv.com")[0], "alice_rp", "alices-dogs-name", WithOperationalState(ActivePrivateState))
+	amActivePub := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice-active-pub.com")[0], "alice_rp", "alices-dogs-name", WithOperationalState(ActivePublicState))
 
 	amId, err := newAuthMethodId()
 	require.NoError(t, err)
 	tests := []struct {
 		name         string
 		in           string
+		opt          []Option
 		want         *AuthMethod
 		wantErrMatch *errors.Template
 	}{
@@ -42,8 +45,26 @@ func TestRepository_LookupAuthMethod(t *testing.T) {
 		},
 		{
 			name: "With existing auth method id",
-			in:   am.GetPublicId(),
-			want: am,
+			in:   amActivePriv.GetPublicId(),
+			want: amActivePriv,
+		},
+		{
+			name: "unauthenticated user - not found using active priv",
+			in:   amActivePriv.GetPublicId(),
+			opt:  []Option{WithUnauthenticatedUser(true)},
+			want: nil,
+		},
+		{
+			name: "unauthenticated user - found active pub",
+			in:   amActivePub.GetPublicId(),
+			opt:  []Option{WithUnauthenticatedUser(true)},
+			want: amActivePub,
+		},
+		{
+			name: "unauthenticated user - found inactive",
+			in:   amInactive.GetPublicId(),
+			opt:  []Option{WithUnauthenticatedUser(true)},
+			want: nil,
 		},
 	}
 
@@ -54,7 +75,7 @@ func TestRepository_LookupAuthMethod(t *testing.T) {
 			repo, err := NewRepository(rw, rw, kmsCache)
 			assert.NoError(err)
 			require.NotNil(repo)
-			got, err := repo.LookupAuthMethod(ctx, tt.in)
+			got, err := repo.LookupAuthMethod(ctx, tt.in, tt.opt...)
 			if tt.wantErrMatch != nil {
 				require.Error(err)
 				assert.Truef(errors.Match(tt.wantErrMatch, err), "want err code: %q got: %q", tt.wantErrMatch, err)
@@ -127,6 +148,139 @@ func TestRepository_ListAuthMethods(t *testing.T) {
 			sort.Slice(got, func(a, b int) bool {
 				return got[a].PublicId < got[b].PublicId
 			})
+			assert.Equal(want, got)
+		})
+	}
+}
+
+func TestRepository_getAuthMethods(t *testing.T) {
+	conn, _ := db.TestSetup(t, "postgres")
+	rw := db.New(conn)
+	wrapper := db.TestWrapper(t)
+	kmsCache := kms.TestKms(t, conn, wrapper)
+	ctx := context.Background()
+
+	tests := []struct {
+		name         string
+		setupFn      func() (authMethodId string, scopeIds []string, want []*AuthMethod)
+		opt          []Option
+		wantErrMatch *errors.Template
+	}{
+		{
+			name: "valid-multi-scopes",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				org, _ := iam.TestScopes(t, iam.TestRepo(t, conn, wrapper))
+				databaseWrapper, err := kmsCache.GetWrapper(context.Background(), org.PublicId, kms.KeyPurposeDatabase)
+				require.NoError(t, err)
+				org2, _ := iam.TestScopes(t, iam.TestRepo(t, conn, wrapper))
+				databaseWrapper2, err := kmsCache.GetWrapper(context.Background(), org2.PublicId, kms.KeyPurposeDatabase)
+				require.NoError(t, err)
+				cert1, _ := testGenerateCA(t, "localhost")
+				cert2, _ := testGenerateCA(t, "127.0.0.1")
+
+				// make a test auth method with all options
+				am1a := TestAuthMethod(
+					t,
+					conn,
+					databaseWrapper,
+					org.PublicId,
+					InactiveState,
+					TestConvertToUrls(t, "https://alice.com")[0],
+					"alice_rp",
+					"alices-dogs-name",
+					WithAudClaims("alice_rp", "alice_rp-2"),
+					WithCallbackUrls(TestConvertToUrls(t, "https://alice.com/callback", "https://alice.com/callback2")...),
+					WithSigningAlgs(RS256, ES256),
+					WithCertificates(cert1, cert2),
+				)
+				am1b := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice.com")[0], "alice_rp-2", "alices-cat-name")
+
+				am2 := TestAuthMethod(t, conn, databaseWrapper2, org2.PublicId, InactiveState, TestConvertToUrls(t, "https://bob.com")[0], "bob_rp", "bobs-dogs-name")
+				return "", []string{am1a.ScopeId, am1b.ScopeId, am2.ScopeId}, []*AuthMethod{am1a, am1b, am2}
+			},
+		},
+		{
+			name: "valid-single-scopes",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				org, _ := iam.TestScopes(t, iam.TestRepo(t, conn, wrapper))
+				databaseWrapper, err := kmsCache.GetWrapper(context.Background(), org.PublicId, kms.KeyPurposeDatabase)
+				require.NoError(t, err)
+				am1a := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice.com")[0], "alice_rp", "alices-dogs-name")
+				am1b := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice.com")[0], "alice_rp-2", "alices-cat-name")
+
+				return "", []string{am1a.ScopeId}, []*AuthMethod{am1a, am1b}
+			},
+		},
+		{
+			name: "valid-auth-method-id",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				org, _ := iam.TestScopes(t, iam.TestRepo(t, conn, wrapper))
+				databaseWrapper, err := kmsCache.GetWrapper(context.Background(), org.PublicId, kms.KeyPurposeDatabase)
+				require.NoError(t, err)
+				am := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice.com")[0], "alice_rp", "alices-dogs-name")
+
+				return am.PublicId, nil, []*AuthMethod{am}
+			},
+		},
+		{
+			name: "with-limits",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				org, _ := iam.TestScopes(t, iam.TestRepo(t, conn, wrapper))
+				databaseWrapper, err := kmsCache.GetWrapper(context.Background(), org.PublicId, kms.KeyPurposeDatabase)
+				require.NoError(t, err)
+				am1a := TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice.com")[0], "alice_rp", "alices-dogs-name")
+				_ = TestAuthMethod(t, conn, databaseWrapper, org.PublicId, InactiveState, TestConvertToUrls(t, "https://alice.com")[0], "alice_rp-2", "alices-cat-name")
+
+				return "", []string{am1a.ScopeId}, []*AuthMethod{am1a}
+			},
+			opt: []Option{WithLimit(1), WithOrder("create_time asc")},
+		},
+		{
+			name: "not-found-auth-method-id",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				return "not-a-valid-id", nil, nil
+			},
+		},
+		{
+			name: "not-found-scope-ids",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				return "", []string{"not-valid-scope-1", "not-valid-scope-2"}, nil
+			},
+		},
+		{
+			name: "no-search-criteria",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				return "", nil, nil
+			},
+			wantErrMatch: errors.T(errors.InvalidParameter),
+		},
+		{
+			name: "search-too-many",
+			setupFn: func() (string, []string, []*AuthMethod) {
+				return "auth-method-id", []string{"scope-id"}, nil
+			},
+			wantErrMatch: errors.T(errors.InvalidParameter),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			assert, require := assert.New(t), require.New(t)
+			r, err := NewRepository(rw, rw, kmsCache)
+			require.NoError(err)
+
+			authMethodId, scopeIds, want := tt.setupFn()
+
+			got, err := r.getAuthMethods(ctx, authMethodId, scopeIds, tt.opt...)
+
+			if tt.wantErrMatch != nil {
+				require.Error(err)
+				assert.Truef(errors.Match(tt.wantErrMatch, err), "want err code: %q got: %q", tt.wantErrMatch, err)
+				assert.Nil(got)
+				return
+			}
+			require.NoError(err)
+			TestSortAuthMethods(t, want)
+			TestSortAuthMethods(t, got)
 			assert.Equal(want, got)
 		})
 	}
